# server/swagger/swagger.yml
basePath: /api/v2
consumes:
    - application/json
definitions:
...
    Board:
        description: Board groups a set of blocks and its layout
        x-go-package: github.com/artipop/xciii/server/model
...
    BoardPatch:
        description: BoardPatch is a patch for modify boards
        x-go-package: github.com/artipop/xciii/server/model
...
    ErrorResponse:
        description: ErrorResponse is an error response
        x-go-package: github.com/artipop/xciii/server/model
    FileUploadResponse:
        description: FileUploadResponse is the response to a file upload
        x-go-package: github.com/artipop/xciii/server/api
...
host: localhost
info:
    contact:
        email: api@focalboard.com
        name: Focalboard
        url: https://www.focalboard.com
    description: Focalboard Server
    license:
        name: Custom
        url: https://github.com/mattermost/focalboard/blob/main/LICENSE.txt
    title: Focalboard Server
    version: 2.0.0
paths:
...
    /boards/{boardID}:
        delete:
            description: Removes a board
            operationId: deleteBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        get:
            description: Returns a board
            operationId: getBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        patch:
            description: Partially updates a board
            operationId: patchBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: board patch to apply
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BoardPatch'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /files/teams/{teamID}/{boardID}/{filename}:
        get:
            description: Returns the contents of an uploaded file
            operationId: getFile
            parameters:
                - description: Team ID
                  in: path
                  name: teamID
                  required: true
                  type: string
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: name of the file
                  in: path
                  name: filename
                  required: true
                  type: string
            produces:
                - application/json
                - image/jpg
                - image/png
                - image/gif
            responses:
                "200":
                    description: success
                "404":
                    description: file not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /teams/{teamID}/boards/{boardID}/files:
        post:
            consumes:
                - multipart/form-data
            description: Upload a binary file, attached to a root block
            operationId: uploadFile
            parameters:
                - description: ID of the team
                  in: path
                  name: teamID
                  required: true
                  type: string
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: The file to upload
                  in: formData
                  name: uploaded file
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/FileUploadResponse'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
produces:
    - application/json
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'Pass session token using Bearer authentication, e.g. set header "Authorization: Bearer <session token>"'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"